requirements.txt
annotated-types>=0.5.0
anyio>=3.6.0
certifi>=2023.7.22
charset-normalizer>=3.0.0
click>=8.0.0
colorama>=0.4.0
cryptography>=41.0.0
fastapi>=0.100.0
h11>=0.14.0
idna>=3.4
python-jose[cryptography]>=3.3.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.0
sniffio>=1.2.0
starlette>=0.27.0
typing_extensions>=4.5.0
typing-inspection>=0.4.0
urllib3>=1.26.0
uvicorn>=0.20.0